handle FE server mechanics in top level interface
Before this I had the actual server be created and handled in each
frontend individually.  I'm assuming all the frontends are going to be
primarily HTTP Rest style things, so it makes sense to abstract that
out.  May need to move it back if I need to rely on things like
websockets or non http protocols in the future.